eXMCP/requirements.txt
mcp
openai
grpcio
grpcio-tools
websockets
fastapi
uvicorn
pydantic
httpx 